Fixed a bug where the entire spectrum was not searched if more than 'top' bins above 'thresh'
Previously the code just ignored any more candidates once the limit was reached.
Now it increases the SNR threshold and restarts the computation.

The code deciding the threshold to use has been moved to slightly earlier in the file
and a wonderful 'goto' statement has been added to jump to just after that decision.